멤버 삭제하는거 수정중

+++ src/main/java/com/crescendo/Util/DeleteScheduler.java
@@ -1,5 +1,7 @@
 package com.crescendo.Util;
 
+import com.crescendo.inquiry.entity.Inquiry;
+import com.crescendo.inquiry.repository.InquiryRepository;
 import com.crescendo.member.repository.MemberRepository;
 import com.crescendo.restore.entity.Restore;
 import com.crescendo.restore.repository.RestoreRepository;
@@ -14,15 +16,31 @@ import java.util.List;
 public class DeleteScheduler {
     private final RestoreRepository restoreRepository;
     private final MemberRepository memberRepository;
-    @Scheduled(fixedRate= 1000 * 60 * 1) // 1분 마다
-    public void performDeleteTask(){
+    private final InquiryRepository inquiryRepository;
+
+    @Scheduled(fixedRate = 1000 * 60 * 1) // 1분 마다
+    public void performDeleteTask() {
         List<Restore> restoreByOverTime = restoreRepository.getRestoreByOverTime();
-        for (Restore restore : restoreByOverTime) {
-            String account = restore.getMember().getAccount();
-            restoreRepository.deleteById(restore.getRestoreNo());
-            memberRepository.deleteById(account);
+        if (restoreByOverTime != null) {
+            for (Restore restore : restoreByOverTime) {
+                String account = restore.getMember().getAccount();
+                changeInquiryMemberToNull(account);
+                restoreRepository.deleteById(restore.getRestoreNo());
+                memberRepository.deleteById(account);
+            }
         }
 
+    }
 
+    // 문의 member null 로 바꾸기
+    private void changeInquiryMemberToNull(String account) {
+        List<Inquiry> allByMemberAccountOrderByInquiryDateTimeDesc = inquiryRepository.findAllByMemberAccountOrderByInquiryDateTimeDesc(account);
+        allByMemberAccountOrderByInquiryDateTimeDesc.forEach(inquiry -> {
+            System.out.println("\n\n\n");
+            System.out.println("inquiry = " + inquiry);
+            System.out.println("\n\n\n");
+            inquiry.setMember(null);
+            System.out.println("inquiry = " + inquiry);
+        });
     }
 }
